dcept_server/requirements.txt
pyiface~=0.0.11
requests~=2.27.1
pyshark~=0.4.5
Flask~=2.1.1
Flask-RESTful~=0.3.9
PyYAML~=6.0